+++ crates/client/storage/src/overrides/schema_v1_override.rs
@@ -5,7 +5,6 @@ use blockifier::execution::contract_class::ContractClass;
 use mp_contract::ContractAbi;
 use mp_storage::{
     PALLET_STARKNET, STARKNET_CONTRACT_ABI, STARKNET_CONTRACT_CLASS, STARKNET_CONTRACT_CLASS_HASH, STARKNET_NONCE,
-    STARKNET_STORAGE,
 };
 use parity_scale_codec::{Decode, Encode};
 // Substrate
@@ -14,8 +13,6 @@ use sp_blockchain::HeaderBackend;
 use sp_runtime::traits::Block as BlockT;
 use sp_storage::StorageKey;
 use starknet_api::api_core::{ClassHash, ContractAddress, Nonce};
-use starknet_api::hash::StarkFelt;
-use starknet_api::state::StorageKey as StarknetStorageKey;
 
 use super::{storage_key_build, storage_prefix_build, StorageOverride};
 
@@ -56,23 +53,6 @@ where
     C: HeaderBackend<B> + StorageProvider<B, BE> + 'static,
     BE: Backend<B> + 'static,
 {
-    fn get_storage_from(
-        &self,
-        block_hash: <B as BlockT>::Hash,
-        address: ContractAddress,
-    ) -> Option<Vec<(StarknetStorageKey, StarkFelt)>> {
-        let storage_storage_prefix = storage_prefix_build(PALLET_STARKNET, STARKNET_STORAGE);
-        let storage = self.query_storage::<Vec<(StarknetStorageKey, StarkFelt)>>(
-            block_hash,
-            &StorageKey(storage_key_build(storage_storage_prefix, &self.encode_storage_key(&address))),
-        );
-
-        match storage {
-            Some(storage) => Some(storage),
-            None => Some(Default::default()),
-        }
-    }
-
     fn contract_class_by_address(
         &self,
         block_hash: <B as BlockT>::Hash,